% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycip.R
\name{mycip}
\alias{mycip}
\title{Confidence interval for p}
\usage{
mycip(x, n, alpha)
}
\arguments{
\item{x}{frequency category}

\item{n}{total frequency}

\item{alpha}{error rate}
}
\value{
list containing ci
}
\description{
Confidence interval for p
}
\examples{
mycip(10,20,0.05)
}
